requirements.txt
boto3==1.18.36
channels-redis==3.3.0
dj-database-url==0.5.0
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-dotenv==1.4.2
django-imagekit==4.0.2
django-storages==1.11.1
django-taggit-templatetags2==1.6.1
gunicorn==20.1.0
pillow==8.3.1
pip-chill==1.0.1
psycopg2-binary==2.9.1
pyopenssl==20.0.1
service-identity==21.1.0
whitenoise==5.3.0
